Remove Fixture.IgnoreCCDWith (#36)
A) The behaviour of `.IgnoreCCDWith` is not fool proof. If a Body has 2
fixtures with categories Cat8 and Cat9 and a another body ignores only
the Cat9, this might result in those two bodies ovelapping (on the
fixture of Cat9) and an overshoot in next Step.

B) If two bodies are known to not be a penetration problem, then the
existing collision mechanism
(CollisionCategories/CollidesWith/CollisionGroup/etc) should be enough
to disable collision during the broadphase.